// frontend/src/api/schemas/doctorDto.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Clic Flow API
 * API documentation for Clic Flow
 * OpenAPI spec version: 1.0
 */

export interface DoctorDto {
  id: string;
  clinicId: string;
  name: string;
  avatarImageUrl?: string;
  /**
   * @minimum 0
   * @maximum 6
   */
  availableFromWeekDay: number;
  /**
   * @minimum 0
   * @maximum 6
   */
  availableToWeekDay: number;
  availableFromTime: string;
  availableToTime: string;
  specialty: string;
  appointmentPriceInCents: number;
  createdAt: string;
  updatedAt: string;
}